refactor(tab-manager): consolidate conversation-handle into chat-state
Merge conversation-handle.svelte.ts into chat-state.svelte.ts, making
createConversationHandle an inner function that closes over orchestrator
state directly. Eliminates the ConversationHandleDeps DI interface which
was over-engineered for a single consumer. Update spec to reflect the
consolidation.
